all menus are functional and pretty
todo: volume slider in main menu is not connected to the volume slider in sample scene. for example when i slide far to the right in the menu and start the game, in this options menu the slider is displayed at its start position. not game breaking but would be nice to fix this